+++ execution/executor-types/src/ledger_update_output.rs
@@ -6,33 +6,25 @@
 use crate::StateComputeResult;
 use anyhow::{ensure, Result};
 use aptos_crypto::HashValue;
-use aptos_storage_interface::cached_state_view::ShardedStateCache;
 use aptos_types::{
     contract_event::ContractEvent,
     epoch_state::EpochState,
-    ledger_info::LedgerInfoWithSignatures,
     proof::accumulator::InMemoryTransactionAccumulator,
-    state_store::{combine_or_add_sharded_state_updates, ShardedStateUpdates},
+    state_store::{combine_or_add_sharded_state_updates, },
     transaction::{
-        block_epilogue::BlockEndInfo, TransactionInfo, TransactionStatus, TransactionToCommit,
+        TransactionInfo,
         Version,
     },
 };
 use itertools::zip_eq;
 use std::sync::Arc;
 
-#[derive(Default, Debug)]
+#[derive(Debug, Default)]
 pub struct LedgerUpdateOutput {
-    pub statuses_for_input_txns: Vec<TransactionStatus>,
-    pub to_commit: Vec<TransactionToCommit>,
-    pub subscribable_events: Vec<ContractEvent>,
+    pub transaction_infos: Vec<TransactionInfo>,
     pub transaction_info_hashes: Vec<HashValue>,
-    pub state_updates_until_last_checkpoint: Option<ShardedStateUpdates>,
-    pub sharded_state_cache: ShardedStateCache,
-    /// The in-memory Merkle Accumulator representing a blockchain state consistent with the
-    /// `state_tree`.
     pub transaction_accumulator: Arc<InMemoryTransactionAccumulator>,
-    pub block_end_info: Option<BlockEndInfo>,
+    pub subscribable_events: Vec<ContractEvent>,
 }
 
 impl LedgerUpdateOutput {
@@ -54,10 +46,6 @@ impl LedgerUpdateOutput {
         &self.transaction_accumulator
     }
 
-    pub fn transactions_to_commit(&self) -> &Vec<TransactionToCommit> {
-        &self.to_commit
-    }
-
     /// Ensure that every block committed by consensus ends with a state checkpoint. That can be
     /// one of the two cases: 1. a reconfiguration (txns in the proposed block after the txn caused
     /// the reconfiguration will be retried) 2. a Transaction::StateCheckpoint at the end of the
@@ -70,66 +58,6 @@ impl LedgerUpdateOutput {
             "Block not ending with a state checkpoint.",
         );
         Ok(())
-    }
-
-    pub fn maybe_select_chunk_ending_ledger_info(
-        &self,
-        verified_target_li: &LedgerInfoWithSignatures,
-        epoch_change_li: Option<&LedgerInfoWithSignatures>,
-        next_epoch_state: Option<&EpochState>,
-    ) -> Result<Option<LedgerInfoWithSignatures>> {
-        if verified_target_li.ledger_info().version() + 1
-            == self.transaction_accumulator.num_leaves()
-        {
-            // If the chunk corresponds to the target LI, the target LI can be added to storage.
-            ensure!(
-                verified_target_li
-                    .ledger_info()
-                    .transaction_accumulator_hash()
-                    == self.transaction_accumulator.root_hash(),
-                "Root hash in target ledger info does not match local computation. {:?} != {:?}",
-                verified_target_li,
-                self.transaction_accumulator,
-            );
-            Ok(Some(verified_target_li.clone()))
-        } else if let Some(epoch_change_li) = epoch_change_li {
-            // If the epoch change LI is present, it must match the version of the chunk:
-
-            // Verify that the given ledger info corresponds to the new accumulator.
-            ensure!(
-                epoch_change_li.ledger_info().transaction_accumulator_hash()
-                    == self.transaction_accumulator.root_hash(),
-                "Root hash of a given epoch LI does not match local computation. {:?} vs {:?}",
-                epoch_change_li,
-                self.transaction_accumulator,
-            );
-            ensure!(
-                epoch_change_li.ledger_info().version() + 1
-                    == self.transaction_accumulator.num_leaves(),
-                "Version of a given epoch LI does not match local computation. {:?} vs {:?}",
-                epoch_change_li,
-                self.transaction_accumulator,
-            );
-            ensure!(
-                epoch_change_li.ledger_info().ends_epoch(),
-                "Epoch change LI does not carry validator set. version:{}",
-                epoch_change_li.ledger_info().version(),
-            );
-            ensure!(
-                epoch_change_li.ledger_info().next_epoch_state() == next_epoch_state,
-                "New validator set of a given epoch LI does not match local computation. {:?} vs {:?}",
-                epoch_change_li.ledger_info().next_epoch_state(),
-                next_epoch_state,
-            );
-            Ok(Some(epoch_change_li.clone()))
-        } else {
-            ensure!(
-                next_epoch_state.is_none(),
-                "End of epoch chunk based on local computation but no EoE LedgerInfo provided. version: {:?}",
-                self.transaction_accumulator.num_leaves().checked_sub(1),
-            );
-            Ok(None)
-        }
     }
 
     pub fn ensure_transaction_infos_match(
@@ -159,6 +87,7 @@ impl LedgerUpdateOutput {
         Ok(())
     }
 
+    /// FIXME(aldenhu): move to upper level
     pub fn as_state_compute_result(
         &self,
         parent_accumulator: &Arc<InMemoryTransactionAccumulator>,
@@ -180,6 +109,7 @@ impl LedgerUpdateOutput {
         )
     }
 
+    /// FIXME(aldenhu): move to upper level
     pub fn combine(&mut self, rhs: Self) {
         assert!(self.block_end_info.is_none());
         assert!(rhs.block_end_info.is_none());
@@ -209,14 +139,17 @@ impl LedgerUpdateOutput {
         self.transaction_accumulator = transaction_accumulator;
     }
 
+    /// FIXME(aldenhu): move to upper level
     pub fn next_version(&self) -> Version {
         self.transaction_accumulator.num_leaves() as Version
     }
 
+    /// FIXME(aldenhu): move to upper level
     pub fn first_version(&self) -> Version {
         self.transaction_accumulator.num_leaves() - self.to_commit.len() as Version
     }
 
+    /// FIXME(aldenhu): move to upper level
     pub fn num_txns(&self) -> usize {
         self.to_commit.len()
     }
