+++ execution/executor/src/components/apply_chunk_output.rs
@@ -6,10 +6,9 @@
 
 use crate::{
     components::{
-        chunk_output::{update_counters_for_processed_chunk, ChunkOutput},
-        in_memory_state_calculator_v2::InMemoryStateCalculatorV2,
+        chunk_output::ChunkOutput, in_memory_state_calculator_v2::InMemoryStateCalculatorV2,
     },
-    metrics::{EXECUTOR_ERRORS, OTHER_TIMERS},
+    metrics::{update_counters_for_processed_chunk, EXECUTOR_ERRORS, OTHER_TIMERS},
 };
 use anyhow::{ensure, Result};
 use aptos_crypto::{hash::CryptoHash, HashValue};
@@ -35,6 +34,7 @@ use aptos_types::{
     },
     write_set::WriteSet,
 };
+use itertools::{izip, Itertools};
 use rayon::prelude::*;
 use std::{iter::repeat, sync::Arc};
 
@@ -42,125 +42,64 @@ pub struct ApplyChunkOutput;
 
 impl ApplyChunkOutput {
     pub fn calculate_state_checkpoint(
-        chunk_output: ChunkOutput,
+        chunk_output: &ChunkOutput,
         parent_state: &StateDelta,
-        append_state_checkpoint_to_block: Option<HashValue>,
         known_state_checkpoints: Option<Vec<Option<HashValue>>>,
         is_block: bool,
-    ) -> Result<(StateDelta, Option<EpochState>, StateCheckpointOutput)> {
-        let ChunkOutput {
-            state_cache,
-            transactions,
-            transaction_outputs,
-            block_end_info,
-        } = chunk_output;
-        let (new_epoch, statuses_for_input_txns, to_commit, to_discard, to_retry) = {
-            let _timer = OTHER_TIMERS.timer_with(&["sort_transactions"]);
-
-            // Separate transactions with different VM statuses, i.e., Keep, Discard and Retry.
-            // Will return transactions with Retry txns sorted after Keep/Discard txns.
-            Self::sort_transactions_with_state_checkpoint(
-                transactions,
-                transaction_outputs,
-                append_state_checkpoint_to_block,
-                block_end_info.clone(),
-            )?
-        };
-
+    ) -> Result<StateCheckpointOutput> {
         // Apply the write set, get the latest state.
-        let (
-            state_updates_vec,
-            state_checkpoint_hashes,
-            result_state,
-            next_epoch_state,
-            state_updates_before_last_checkpoint,
-            sharded_state_cache,
-        ) = {
-            let _timer = OTHER_TIMERS
-                .with_label_values(&["calculate_for_transactions"])
-                .start_timer();
-            InMemoryStateCalculatorV2::calculate_for_transactions(
-                parent_state,
-                state_cache,
-                &to_commit,
-                new_epoch,
-                is_block,
-            )?
-        };
-
-        let mut state_checkpoint_output = StateCheckpointOutput::new(
-            TransactionsByStatus::new(statuses_for_input_txns, to_commit, to_discard, to_retry),
-            state_updates_vec,
-            state_checkpoint_hashes,
-            state_updates_before_last_checkpoint,
-            sharded_state_cache,
-            block_end_info,
-        );
+        let mut res = InMemoryStateCalculatorV2::calculate_for_transactions(
+            parent_state,
+            chunk_output,
+            is_block,
+        )?;
 
         // On state sync/replay, we generate state checkpoints only periodically, for the
         // last state checkpoint of each chunk.
         // A mismatch in the SMT will be detected at that occasion too. Here we just copy
         // in the state root from the TxnInfo in the proof.
         if let Some(state_checkpoint_hashes) = known_state_checkpoints {
-            state_checkpoint_output
-                .check_and_update_state_checkpoint_hashes(state_checkpoint_hashes)?;
+            res.check_and_update_state_checkpoint_hashes(state_checkpoint_hashes)?;
         }
 
-        Ok((result_state, next_epoch_state, state_checkpoint_output))
+        Ok(res)
     }
 
     pub fn calculate_ledger_update(
-        state_checkpoint_output: StateCheckpointOutput,
-        base_txn_accumulator: Arc<InMemoryTransactionAccumulator>,
-    ) -> Result<(LedgerUpdateOutput, Vec<Transaction>, Vec<Transaction>)> {
-        let (
-            txns,
-            state_updates_vec,
-            state_checkpoint_hashes,
-            state_updates_before_last_checkpoint,
-            sharded_state_cache,
-            block_end_info,
-        ) = state_checkpoint_output.into_inner();
-
-        let (statuses_for_input_txns, to_commit, to_discard, to_retry) = txns.into_inner();
-
-        update_counters_for_processed_chunk(
-            to_commit.txns(),
-            to_commit.parsed_outputs(),
-            "execution",
-        );
-        update_counters_for_processed_chunk(
-            to_discard.txns(),
-            to_discard.parsed_outputs(),
-            "execution",
-        );
-        update_counters_for_processed_chunk(
-            to_retry.txns(),
-            to_retry.parsed_outputs(),
-            "execution",
-        );
-
-        // Calculate TransactionData and TransactionInfo, i.e. the ledger history diff.
+        chunk_output: &ChunkOutput,
+        state_checkpoint_hashes: &[Option<HashValue>],
+        base_txn_accumulator: &InMemoryTransactionAccumulator,
+    ) -> Result<LedgerUpdateOutput> {
         let _timer = OTHER_TIMERS.timer_with(&["assemble_ledger_diff_for_block"]);
 
-        let (txns_to_commit, transaction_info_hashes, subscribable_events) =
-            Self::assemble_ledger_diff(to_commit, state_updates_vec, state_checkpoint_hashes);
+        // Update counters.
+        chunk_output.update_counters_for_processed_chunk();
+
+        // Calculate hashes
+        let to_commit = &chunk_output.to_commit;
+        let txn_outs = to_commit.parsed_outputs();
+
+        let (event_hashes, writeset_hashes) = Self::calculate_events_and_writeset_hashes(txn_outs);
+
+        // Assemble `TransactionInfo`s
+        let (transaction_infos, subscribable_events) = Self::assemble_transaction_infos(
+            &to_commit,
+            &state_checkpoint_hashes,
+            &event_hashes,
+            &writeset_hashes,
+        );
+
+        // Calculate root hash
+        let transaction_info_hashes = transaction_infos.iter().map(CryptoHash::hash).collect_vec();
         let transaction_accumulator =
             Arc::new(base_txn_accumulator.append(&transaction_info_hashes));
-        Ok((
-            LedgerUpdateOutput {
-                statuses_for_input_txns,
-                to_commit: txns_to_commit,
-                subscribable_events,
-                transaction_info_hashes,
-                state_updates_until_last_checkpoint: state_updates_before_last_checkpoint,
-                sharded_state_cache,
-                transaction_accumulator,
-                block_end_info,
-            },
-            to_discard.into_txns(),
-            to_retry.into_txns(),
-        ))
+
+        Ok(LedgerUpdateOutput {
+            transaction_infos,
+            transaction_info_hashes,
+            transaction_accumulator,
+            subscribable_events,
+        })
     }
 
     pub fn apply_chunk(
@@ -312,106 +251,64 @@ impl ApplyChunkOutput {
         ))
     }
 
-    fn assemble_ledger_diff(
-        to_commit_from_execution: TransactionsWithParsedOutput,
-        state_updates_vec: Vec<ShardedStateUpdates>,
-        state_checkpoint_hashes: Vec<Option<HashValue>>,
-    ) -> (Vec<TransactionToCommit>, Vec<HashValue>, Vec<ContractEvent>) {
-        // these are guaranteed by caller side logic
-        assert_eq!(to_commit_from_execution.len(), state_updates_vec.len());
-        assert_eq!(
-            to_commit_from_execution.len(),
-            state_checkpoint_hashes.len()
-        );
-
-        let num_txns = to_commit_from_execution.len();
-        let mut to_commit = Vec::with_capacity(num_txns);
-        let mut txn_info_hashes = Vec::with_capacity(num_txns);
-        let hashes_vec =
-            Self::calculate_events_and_writeset_hashes(to_commit_from_execution.parsed_outputs());
-
+    fn assemble_transaction_infos(
+        to_commit: &TransactionsWithParsedOutput,
+        state_checkpoint_hashes: &[Option<HashValue>],
+        event_hashes: &[HashValue],
+        writeset_hashes: &[HashValue],
+    ) -> (Vec<TransactionInfo>, Vec<ContractEvent>) {
         let _timer = OTHER_TIMERS.timer_with(&["process_events_and_writeset_hashes"]);
-        let hashes_vec: Vec<(HashValue, HashValue)> = hashes_vec
-            .into_par_iter()
-            .map(|(event_hashes, write_set_hash)| {
-                (
-                    InMemoryEventAccumulator::from_leaves(&event_hashes).root_hash(),
-                    write_set_hash,
-                )
-            })
-            .collect();
 
-        let mut all_subscribable_events = Vec::new();
-        let (to_commit_txns, to_commit_outputs) = to_commit_from_execution.into_inner();
-        for (
-            txn,
-            txn_output,
-            state_checkpoint_hash,
-            state_updates,
-            (event_root_hash, write_set_hash),
-        ) in itertools::izip!(
-            to_commit_txns,
-            to_commit_outputs,
+        izip!(
+            to_commit.iter(),
             state_checkpoint_hashes,
-            state_updates_vec,
-            hashes_vec
-        ) {
-            let (write_set, events, per_txn_reconfig_events, gas_used, status, auxiliary_data) =
-                txn_output.unpack();
-
-            let subscribable_events: Vec<ContractEvent> = events
-                .iter()
-                .filter(|evt| should_forward_to_subscription_service(evt))
-                .cloned()
-                .collect();
-            let txn_info = match &status {
-                TransactionStatus::Keep(status) => TransactionInfo::new(
+            event_hashes,
+            writeset_hashes
+        )
+        .map(
+            |((txn, txn_out), state_checkpoint_hash, event_root_hash, write_set_hash)| {
+                let subscribable_events: Vec<ContractEvent> = txn_out
+                    .events()
+                    .iter()
+                    .filter(should_forward_to_subscription_service)
+                    .cloned()
+                    .collect();
+                let txn_info = TransactionInfo::new(
                     txn.hash(),
                     write_set_hash,
                     event_root_hash,
-                    state_checkpoint_hash,
-                    gas_used,
-                    status.clone(),
-                ),
-                _ => unreachable!("Transaction sorted by status already."),
-            };
-            let txn_info_hash = txn_info.hash();
-            txn_info_hashes.push(txn_info_hash);
-            let txn_to_commit = TransactionToCommit::new(
-                txn,
-                txn_info,
-                state_updates,
-                write_set,
-                events,
-                !per_txn_reconfig_events.is_empty(),
-                auxiliary_data,
-            );
-            all_subscribable_events.extend(subscribable_events);
-            to_commit.push(txn_to_commit);
-        }
-        (to_commit, txn_info_hashes, all_subscribable_events)
+                    state_checkpoint_hash.cloned(),
+                    txn_out.gas_used(),
+                    txn_out.status().as_kept_status().expect("Already sorted."),
+                );
+                (txn_info, subscribable_events)
+            },
+        )
+        .unzip()
     }
 
     fn calculate_events_and_writeset_hashes(
-        to_commit_from_execution: &[ParsedTransactionOutput],
-    ) -> Vec<(Vec<HashValue>, HashValue)> {
+        to_commit: &[ParsedTransactionOutput],
+    ) -> (Vec<HashValue>, Vec<HashValue>) {
         let _timer = OTHER_TIMERS.timer_with(&["calculate_events_and_writeset_hashes"]);
 
-        let num_txns = to_commit_from_execution.len();
-        to_commit_from_execution
+        let num_txns = to_commit.len();
+        to_commit
             .par_iter()
             .with_min_len(optimal_min_len(num_txns, 64))
             .map(|txn_output| {
+                let event_hashes = txn_output
+                    .events()
+                    .iter()
+                    .map(CryptoHash::hash)
+                    .collect::<Vec<_>>();
+
                 (
-                    txn_output
-                        .events()
-                        .iter()
-                        .map(CryptoHash::hash)
-                        .collect::<Vec<_>>(),
+                    InMemoryEventAccumulator::from_leaves(&event_hashes).root_hash(),
                     CryptoHash::hash(txn_output.write_set()),
                 )
             })
-            .collect::<Vec<_>>()
+            .unzip()
     }
 }
 
@@ -460,7 +357,7 @@ fn assemble_ledger_diff_should_filter_subscribable_events() {
         ShardedStateUpdates::default(),
     ];
     let state_checkpoint_hashes = vec![Some(HashValue::zero()), Some(HashValue::zero())];
-    let (_, _, subscribable_events) = ApplyChunkOutput::assemble_ledger_diff(
+    let (_, _, subscribable_events) = ApplyChunkOutput::calculate_transaction_infos(
         txns_n_outputs,
         state_updates_vec,
         state_checkpoint_hashes,
