+++ execution/executor/src/components/mod.rs
@@ -10,6 +10,7 @@ pub mod chunk_commit_queue;
 pub mod chunk_output;
 pub mod in_memory_state_calculator_v2;
 
-pub mod transaction_chunk;
 mod chunk_proof;
 mod speculative_state;
+pub mod transaction_chunk;
+mod make_chunk_output;
