+++ execution/executor/src/fuzzing.rs
@@ -75,7 +75,12 @@ impl TransactionBlockExecutor for FakeVM {
         onchain_config: BlockExecutorConfigFromOnchain,
         append_state_checkpoint_to_block: Option<HashValue>,
     ) -> Result<ChunkOutput> {
-        ChunkOutput::by_transaction_execution::<FakeVM>(transactions, state_view, onchain_config, append_state_checkpoint_to_block)
+        ChunkOutput::by_transaction_execution::<FakeVM>(
+            transactions,
+            state_view,
+            onchain_config,
+            append_state_checkpoint_to_block,
+        )
     }
 }
