+++ execution/executor/src/mock_vm/mod.rs
@@ -67,7 +67,12 @@ impl TransactionBlockExecutor for MockVM {
         onchain_config: BlockExecutorConfigFromOnchain,
         append_state_checkpoint_to_block: Option<HashValue>,
     ) -> Result<ChunkOutput> {
-        ChunkOutput::by_transaction_execution::<MockVM>(transactions, state_view, onchain_config, append_state_checkpoint_to_block)
+        ChunkOutput::by_transaction_execution::<MockVM>(
+            transactions,
+            state_view,
+            onchain_config,
+            append_state_checkpoint_to_block,
+        )
     }
 }
