+++ storage/backup/backup-cli/src/backup_types/transaction/restore.rs
@@ -629,8 +629,7 @@ impl TransactionRestoreBatchController {
                         .with_label_values(&["commit_txn_chunk"])
                         .start_timer();
                     tokio::task::spawn_blocking(move || {
-                        let committed_chunk = chunk_replayer.commit()?;
-                        let v = committed_chunk.result_state.current_version.unwrap_or(0);
+                        let v = chunk_replayer.commit()?;
                         let total_replayed = v - first_version + 1;
                         TRANSACTION_REPLAY_VERSION.set(v as i64);
                         info!(
